backend/requirements.txt
fastapi
uvicorn[standard]
SQLModel
alembic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
sqlmodel
pydantic<2
python-multipart